requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.15.0
azure-search-documents==11.4.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
dataclasses-json==0.6.4
diskcache==5.6.3
distro==1.9.0
fairscale==0.4.13
fastapi==0.110.1
filelock==3.13.3
fire==0.5.0
frozenlist==1.4.1
fsspec==2024.3.1
h11==0.14.0
hiq-python==1.1.12
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
isodate==0.6.1
itree==0.0.5
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain-community==0.0.31
langchain-core==0.1.40
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.1.40
llama_cpp_python==0.2.47
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
mpmath==1.3.0
msal==1.28.0
msal-extensions==1.1.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
openai==1.16.2
orjson==3.10.0
packaging==23.2
portalocker==2.8.2
psutil==5.9.8
py-itree==0.0.19
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
pyllama==0.0.9
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
safetensors==0.4.2
sentencepiece==0.1.97
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
sympy==1.12
tenacity==8.2.3
termcolor==2.4.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.2
transformers==4.39.3
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
